% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bookshop_orders}
\alias{bookshop_orders}
\title{Antiquarian Bookshop Orders}
\format{
A data frame with columns:
\describe{
  \item{id}{ID of the order.}
  \item{datetime}{Date and time of the order.}
  \item{quantity}{Number of purchased books.}
  \item{duration}{Number of minutes since the last order.}
  \item{duration_adj}{Duration since the last order adjusted for diurnal pattern.}
}
}
\usage{
bookshop_orders
}
\description{
Individual orders of a Czech antiquarian bookshop from June 8, 2018 to December 20, 2018.
This dataset is analyzed in Tomanová and Holý (2021).
}
\references{
Tomanová, P. and Holý, V. (2021). Clustering of Arrivals in Queueing Systems: Autoregressive Conditional Duration Approach. \emph{Central European Journal of Operations Research}, \strong{29}(3), 859–874. \doi{10.1007/s10100-021-00744-7}.
}
\keyword{datasets}
