% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bookshop_sales}
\alias{bookshop_sales}
\title{Antiquarian Bookshop Sales}
\format{
A data frame with columns:
\describe{
  \item{order}{ID of the order.}
  \item{time}{Time of the order.}
  \item{quantity}{Number of purchased books. Zero value means the order was canceled.}
}
}
\source{
Petra Tomanová (\email{petra.tomanova@vse.cz}).
}
\usage{
bookshop_sales
}
\description{
Individual orders of a Czech antiquarian bookshop from June 8, 2018 to December 20, 2018.
This dataset is analyzed in Tomanová and Holý (2021).
Details on the bookshop can be also found in Tomanová and Černý (2022).
}
\references{
Tomanová, P. and Černý, M. (2022). Efficiency of Antiquarian Bookshops in Informationally Complete Markets. \emph{Central European Journal of Operations Research}, \strong{30}(2), 573–593. doi: \href{https://doi.org/10.1007/s10100-021-00780-3}{10.1007/s10100-021-00780-3}.

Tomanová, P. and Holý, V. (2021). Clustering of Arrivals in Queueing Systems: Autoregressive Conditional Duration Approach. \emph{Central European Journal of Operations Research}, \strong{29}(3), 859–874. doi: \href{https://doi.org/10.1007/s10100-021-00744-7}{10.1007/s10100-021-00744-7}.
}
\keyword{datasets}
