% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_car_market_cap}
\alias{german_car_market_cap}
\title{Market Capitalization of German Car Manufacturers}
\format{
A data frame with columns:
\describe{
  \item{year}{Year.}
  \item{car_manufacturer}{Car manufacturer.}
  \item{market_cap}{Market capitalization in billions of euros.}
}
}
\source{
Thomson Reuters (\href{https://www.thomsonreuters.com/}{www.thomsonreuters.com}).
}
\usage{
german_car_market_cap
}
\description{
Market capitalization of the "Germany's Big Three" automobile manufacturers – Volkswagen Group, Mercedes-Benz Group, and BMW.
Market capitalization is reported in billions of euros and covers the period 1994–2021.
}
\keyword{datasets}
