% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_hockey_championships}
\alias{ice_hockey_championships}
\title{Results of the Ice Hockey World Championships}
\format{
A list with components:
\describe{
  \item{rankings}{A matrix of final rankings. Rows correspond to years, columns to teams. Value \code{Inf} means that the team did not advance to the championship. Value \code{NA} means that the championship did not take place.}
  \item{hosts}{A matrix of dummy variables indicating whether the team hosted the championship. Rows correspond to years, columns to teams. Multiple hosts of one championship is possible. Value \code{NA} means that the championship did not take place.}
}
}
\source{
International Ice Hockey Federation (\href{https://www.iihf.com/}{www.iihf.com}).
}
\usage{
ice_hockey_championships
}
\description{
The dataset contains the results of the annual men's Ice Hockey World Championships from 1998 to 2021.
In 1998, the International Ice Hockey Federation set the number of teams participating in the championships at 16.
Since 1998, a total of 24 teams have qualified for the championship division.
This dataset is analyzed in Holý and Zouhar (2022).
}
\references{
Holý, V. and Zouhar, J. (2022). Modelling Time-Varying Rankings with Autoregressive and Score-Driven Dynamics. Journal of the Royal Statistical Society: Series C (Applied Statistics), \strong{71}(5). \doi{10.1111/rssc.12584}.
}
\keyword{datasets}
