% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500_daily}
\alias{sp500_daily}
\title{Results of the Ice Hockey World Championships}
\format{
A data frame with columns:
\describe{
  \item{date}{Trading day.}
  \item{open}{Opening price of the day.}
  \item{high}{Highest price of the day.}
  \item{low}{Lowest price of the day.}
  \item{close}{Closing price of the day.}
}
}
\source{
Nasdaq (\href{https://www.nasdaq.com/market-activity/index/spx}{www.nasdaq.com/market-activity/index/spx}).
}
\usage{
sp500_daily
}
\description{
Daily opening, highest, lowest, and closing prices of the Standard and Poor's 500 stock market index (SPX) from 2013.
}
\keyword{datasets}
