% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toilet_paper_sales}
\alias{toilet_paper_sales}
\title{Daily Toilet Paper Sales}
\format{
A data frame with columns:
\describe{
  \item{date}{Date.}
  \item{monday}{Dummy variable indicating whether it is Monday.}
  \item{tuesday}{Dummy variable indicating whether it is Tuesday.}
  \item{wednesday}{Dummy variable indicating whether it is Wednesday.}
  \item{thursday}{Dummy variable indicating whether it is Thursday.}
  \item{friday}{Dummy variable indicating whether it is Friday.}
  \item{saturday}{Dummy variable indicating whether it is Saturday.}
  \item{sunday}{Dummy variable indicating whether it is Sunday.}
  \item{promo}{Dummy variable indicating whether the product is promoted.}
  \item{quantity}{Number of packs sold.}
}
}
\usage{
toilet_paper_sales
}
\description{
The daily number of toilet paper packs sold in a European store in 2001 and 2002.
The \code{promo} variable indicates whether the product was promoted in a campaign.
Missing values correspond to the days when the store was closed.
}
\keyword{datasets}
